Refactor Stops Layer
